+++ inst/app/global.R
@@ -8,6 +8,8 @@ ifelse (grepl("radiant.multivariate", getwd()) && file.exists("../../inst") , ".
 ## setting path for figures in help files
 addResourcePath("figures_multivariate", "tools/help/figures/")
 
+if (is.null(getOption("radiant.path.model"))) options(radiant.path.model = system.file(package = "radiant.model"))
+
 ## loading urls and ui
 source("init.R", encoding = getOption("radiant.encoding"), local = TRUE)
 options(radiant.url.patterns = make_url_patterns())