+++ NEWS.md
@@ -1,7 +1,12 @@
-# CHANGES IN radiant.multivariate VERSION 0.6.1 (unreleased)
+# CHANGES IN radiant.multivariate VERSION 0.6.3 (unreleased)
 
 ## NEW FEATURES
 - Added k-medians, from the Gmedians package, as an option in _Multivariate > K-clustering_
+- Added additional rotation options in _Multivariate > Factor_
+- Added predict tab for conjoint
+- Added option to analyse conjoint data for multiple respondents (`By`) or for a specific respondent (`Show`)
+- Store PWs or IWs from conjoint analysis for multiple respondents
+- Derive and store predictions based on conjoint analysis for multiple respondents
 
 ## BUG FIXES